Cast random() properly and fix comment

+++ grid.c
@@ -136,12 +136,12 @@ void grid_set(grid_t *grid, coord_t location, square_t square) {
  */
 void grid_add_bombs(grid_t *grid, unsigned int bombs) {
 	/* XXX: don't need to check for duplicates because the period of
-	 * random(2) is 8x bigger than UINT_MAX-1, the maximum number of bombs
+	 * random(3) is 8x bigger than UINT_MAX-1, the maximum number of bombs
 	 * that can be placed. This IS platform dependant! If an unsigned int is
 	 * any larger than 34 bits, duplicate checking will need to be done.
 	 */
 	for (/* EMPTY */; bombs != 0; bombs--) {
-		grid_set(grid, coord_new(random() % grid->width, random() % grid->height), SQUARE_BOMB);
+		grid_set(grid, coord_new((unsigned)random() % grid->width, (unsigned)random() % grid->height), SQUARE_BOMB);
 	}
 }
